+++ app/src/main/java/com/goo32v2/cooldict/words/NavDrawerFragment.java
@@ -3,10 +3,13 @@ package com.goo32v2.cooldict.words;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
+import android.support.v7.app.ActionBarDrawerToggle;
+import android.support.v7.widget.Toolbar;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.goo32v2.cooldict.R;
 import com.goo32v2.cooldict.data.models.DictionaryModel;
 
 import java.util.List;
@@ -42,7 +45,20 @@ public class NavDrawerFragment extends Fragment {
         super.onDetach();
     }
 
-    public void setupNavigationDrawer(List<DictionaryModel> items){
+    private void setupDrawer(Toolbar toolbar){
+        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
+                this,
+                mDrawerLayout,
+                toolbar,
+                R.string.navigation_drawer_open,
+                R.string.navigation_drawer_close
+        );
+        mDrawerLayout.addDrawerListener(toggle);
+        toggle.syncState();
+    }
+
+    //use for first setup
+    public void setupNavigationDrawer(Toolbar toolbar){
         //        CallbackHelper callback = new CallbackHelper();
 //        mDictionaryRepository.get(callback);
 //
@@ -64,7 +80,7 @@ public class NavDrawerFragment extends Fragment {
 
     }
 
-    public void update(){
+    public void update(List<DictionaryModel> entries){
 
     }
